pyproject.toml
[project]
name = "datapersistence"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.110.0",
    "httpx>=0.25.0",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.3.3",
    "pydantic>=2.6.0",
    "pytest>=9.0.3",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.23",
    "utils>=1.0.2",
    "uuid7>=0.1.0",
    "uvicorn>=0.24.0",
]
